Refactor layout and styling of components; replace unused elements with a responsive navigation bar and footer; update color scheme for improved visual consistency.
