+++ app/footer/page.tsx
@@ -1,0 +1,5 @@
+import Footer from "@/components/footer";
+
+export default function FooterPage() {
+  return <Footer />;
+}
